fix(ios): fix notification crash when debug=true
At the beginning of `onFinishLaunching`, `center.delegate` is set to
`self`. Later on, when debugging is enabled, the current value of
`center.delegate` is saved info `prevNotificationDelegate`. Since this
has already been set to `self`, a recursion and later a stack overflow
happens.

The commit that added the code
  80be0fd ("iOS avoid taking control of UNUserNotificationCenter", 2018-08-20)
has not removed the block at the beginning, but the original one referenced
there did:
  https://github.com/mauron85/react-native-background-geolocation/pull/268/files

So that was probably a mis-merge.

Fixes #565